wasm: table and elem edges should go table -> elem
It was the other way around because at the binary level the element references
which table it is populating, but logically it is the table that owns its
elements. It gives us much more useful data this way: we can see when the table
is heavy in the dominators tree, because a bunch of dynamic indirect calls that
the linker couldn't remove got entrenched in the binary.