Switch to `wasmparser` from `parity-wasm`
This commit switches the wasm size analysis from the `parity-wasm` crate
to the `wasmparser` crate. The primary downside of `parity-wasm` is that
it doesn't retain offsets into the original file, so the current way
used to see how big a section is on disk is to serialize it and measure
the serialized size. The serialized size, however, is not always equal
to the original size when accounting for differences like leb128
encodings.

The `wasmparser` crate has offset information plumbed throughout
parsing, allowing us to faithfully represent the size of various
constructs losslessly from the original file. The `wasmparser` crate,
unlike `parity-wasm`, doesn't provide an IR to work with. This means
that we have to do a bit more data munging ourselves in a few locations.