Attempt to account for 100% of bytes in wasm files
This commit attempts to ensure that twiggy is aware of 100% of the bytes
in a wasm file, ensuring that no byte escapes twiggy's watchful gaze.
The main source of bytes unaccounted for was headers in each section,
for example the byte indicating what section it is as well as the bytes
used to describe how large the section is. A few other minor sources of
"bloat", like the 8 magic wasm bytes, were also unaccounted for.

This change should ensure that all these bytes are accounted for one way
or another in a wasm file, ideally making twiggy a lossless
representation of a wasm file!

Closes #248